// src/Enums/StoreEnums.ts
export enum CharactersStore {
    MODULE_NAME = 'charactersModule',
    STATE_CHARACTERS = 'characters',
    SET_CHARACTERS = 'setCharacters'
}

export enum WeaponsStore {
    MODULE_NAME = 'weaponsModule',
    STATE_WEAPONS = 'weapons',
    SET_WEAPONS = 'setWeapons'
}

export enum EventWishStore {
    MODULE_NAME = 'eventWish',
    STATE_TOTAL_WISHES = 'totalWishes',
    SET_TOTAL_WISHES = 'setEventWishes'
}

export enum StandartWishStore {
    MODULE_NAME = 'standartWish',
    STATE_TOTAL_WISHES = 'totalWishes',
    SET_TOTAL_WISHES = 'setStandartWishes'
}
export enum EventWeaponWishStore {
    MODULE_NAME = 'eventWeaponWish',
    STATE_TOTAL_WISHES = 'totalWishes',
    SET_TOTAL_WISHES = 'setEventWeaponWishes'
}

export enum PrimogemsStore {
    MODULE_NAME = 'primogemsModule',
    STATE_PRIMOGEMS = 'primogems',
    SET_PRIMOGEMS = 'setPrimogems'
}

export enum BannerStore {
    MODULE_NAME = 'bannerModule',
    STATE_CURRENT_BANNER = 'banner',
    SET_CURRENT_BANNER = 'setBanner'
}

export enum WishListStore {
    MODULE_NAME = 'wishListModule',
    STATE_WISH_LIST = 'wishList',
    SET_WISH_LIST = 'setWishList'
}